// apps/realtime-archiver/src/services/analytics-types.ts
export type AnalyticsTradeRecord = {
	dt: string;
	ts: number;
	action: string;
	actionexplanation?: string;
	marketindex: number;
	market: string;
	markettype: string;
	filler?: string;
	fillerreward?: number;
	fillrecordid: number;
	baseassetamountfilled: number;
	quoteassetamountfilled: number;
	takerfee: number;
	makerfee: number;
	referrerreward: number;
	quoteassetamountsurplus: number;
	spotfulfillmentmethodfee: number;
	taker?: string;
	taker_authority?: string;
	takerorderid: number;
	takerorderdirection?: string;
	takerorderbaseassetamount: number;
	takerordercumulativebaseassetamountfilled: number;
	takerordercumulativequoteassetamountfilled: number;
	maker?: string;
	maker_authority?: string;
	makerorderid: number;
	makerorderdirection?: string;
	makerorderbaseassetamount: number;
	makerordercumulativebaseassetamountfilled: number;
	makerordercumulativequoteassetamountfilled: number;
	oracleprice: number;
	txsig: string;
	slot: number;
	eventtype: 'TradeRecord';
	txsigindex: number;
	source: string;
	marketfilter?: string;
	bitflags: number;
	takerexistingquoteentryamount: number;
	takerexistingbaseassetamount: number;
	makerexistingquoteentryamount: number;
	makerexistingbaseassetamount: number;
	triggerprice: number;
	builderidx?: string;
	builderfee: number;
	year: string;
	month: string;
	day: string;
};

export type AnalyticsDepositRecord = {
	dt: string;
	ts: number;
	userauthority: string;
	user: string;
	direction: string;
	depositrecordid: number;
	amount: number;
	marketindex: number;
	market: string;
	oracleprice: number;
	marketdepositbalance: number;
	marketwithdrawbalance: number;
	marketcumulativedepositinterest: number;
	marketcumulativeborrowinterest: number;
	totaldepositsafter: number;
	totalwithdrawsafter: number;
	explanation: string;
	transferuser?: string;
	txsig: string;
	slot: number;
	eventtype: 'DepositRecord';
	txsigindex: number;
	source: string;
	signer?: string;
	usertokenamountafter: number;
	year: string;
	month: string;
	day: string;
};